refactor React components into own files
This will make each component easier to maintain and allows a more
modular design.

This is done with: ObjectsList, SideBar, Path, BrowserUpdate, and
ConfirmModal, all pre-existing components.
